refactor(1단계 리팩터링): 패키지 구조 변경
