+++ app/src/main/java/woowacourse/shopping/data/local/WoowaShoppingContract.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data
+package woowacourse.shopping.data.local
 
 import android.provider.BaseColumns
