+++ app/src/main/java/woowacourse/shopping/data/mapper/ProductMapper.kt
@@ -1,7 +1,7 @@
 package woowacourse.shopping.data.mapper
 
-import woowacourse.shopping.data.dto.ProductDto
-import woowacourse.shopping.data.dto.ProductListDto
+import woowacourse.shopping.data.remote.dto.ProductDto
+import woowacourse.shopping.data.remote.dto.ProductListDto
 import woowacourse.shopping.domain.model.CartItem
 import woowacourse.shopping.domain.model.CartProduct
