+++ app/src/main/java/woowacourse/shopping/data/mapper/ShoppingCartMapper.kt
@@ -1,6 +1,6 @@
 package woowacourse.shopping.data.mapper
 
-import woowacourse.shopping.data.dto.ShoppingCartDto
+import woowacourse.shopping.data.remote.dto.ShoppingCartDto
 import woowacourse.shopping.domain.model.CartItem
 import woowacourse.shopping.domain.model.CartProduct
