+++ app/src/main/java/woowacourse/shopping/data/remote/dto/ProductListDto.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.dto
+package woowacourse.shopping.data.remote.dto
 
 data class ProductListDto(
     val products: List<ProductDto>,
